Re-align tests with the rest of the project structure
